x86/resctrl: Claim get_{mon,ctrl}_domain_from_cpu() helpers for resctrl
get_{mon,ctrl}_domain_from_cpu() are handy helpers that both the arch
code and resctrl need to use. Rename them to have a resctrl_ prefix
and move them to a header file.

Tested-by: Carl Worth <carl@os.amperecomputing.com> # arm64
Tested-by: Shaopeng Tan <tan.shaopeng@jp.fujitsu.com>
---
Changes from v5:
 * Added the word from to a comment.
(cherry picked from commit f9a611dff7308edb5c1827ef5b255ff25d43c328 https://web.git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git/log/?h=mpam/snapshot/v6.14-rc1)
